requirements.txt
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cx-Oracle==8.3.0
Flask==2.2.5
Flask-Login==0.6.3
Flask-Mercadopago==1.0.0
Flask-SQLAlchemy==3.0.5
greenlet==3.1.1
gunicorn==23.0.0
idna==3.10
importlib-metadata==6.7.0
itsdangerous==2.1.2
Jinja2==3.1.5
MarkupSafe==2.1.5
mercadopago==2.2.3
mysql-connector-python==8.0.33
packaging==24.0
Pillow==9.5.0
protobuf==3.20.3
psycopg2==2.9.9
psycopg2-binary==2.9.9
PyMySQL==1.1.1
pypng==0.20220715.0
python-dotenv==0.21.1
qrcode==7.4.2
requests==2.31.0
SQLAlchemy==2.0.38
typing_extensions==4.7.1
urllib3==2.0.7
Werkzeug==2.2.3
zipp==3.15.0
